refactor(mutator set): rename active window to swbf suffix
"Active window" is an overloaded term, referring to two different
things, see #71 [1]. The rename clarifies that this struct refers
to the portion of the SWBF that is represented explicity.
Additionally, a generic type parameter is introduced which
determines the size of the suffix. (In future, this size may be
different from the window size.)

[1]: https://github.com/Neptune-Crypto/neptune-core/issues/71

+++ src/models/blockchain/transaction/validity/tasm/compute_indices.rs
@@ -60,7 +60,7 @@ impl BasicSnippet for ComputeIndices {
             data_type: DataType::U64,
         }));
         let get_swbf_indices = library.import(Box::new(GetSwbfIndices {
-            window_size: WINDOW_SIZE,
+            window_size: WINDOW_SIZE as u32,
             num_trials: NUM_TRIALS as usize,
         }));
 
@@ -184,7 +184,7 @@ impl Function for ComputeIndices {
         stack.push(item.values()[1]);
         stack.push(item.values()[0]);
         let get_swbf_indices = GetSwbfIndices {
-            window_size: WINDOW_SIZE,
+            window_size: WINDOW_SIZE as u32,
             num_trials: NUM_TRIALS as usize,
         };
         get_swbf_indices.rust_shadow(stack, memory);
