+++ src/util_types/mutator_set/shared.rs
@@ -10,7 +10,7 @@ use twenty_first::util_types::mmr::mmr_membership_proof::MmrMembershipProof;
 use super::chunk_dictionary::ChunkDictionary;
 use super::removal_record::RemovalRecord;
 
-pub const WINDOW_SIZE: u32 = 1 << 20;
+pub const WINDOW_SIZE: usize = 1 << 20;
 pub const CHUNK_SIZE: u32 = 1 << 12;
 pub const BATCH_SIZE: u32 = 1 << 3;
 pub const NUM_TRIALS: u32 = 45;
